Finish ApplicationUser Management

+++ BTS.Web/Controllers/ApplicationGroupController.cs
@@ -10,28 +10,31 @@ using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
 using BTS.Web.Infrastructure.Extensions;
+using BTS.Data.ApplicationModels;
+using BTS.Common;
 
 namespace BTS.Web.Controllers
 {
+    [AuthorizeRoles(CommonConstants.System_CanView_Role)]
     public class ApplicationGroupController : BaseController
     {
         private IApplicationGroupService _appGroupService;
-        private IApplicationRoleService _appRoleService;
         private ApplicationUserManager _userManager;
 
         public ApplicationGroupController(IErrorService errorService,
-            IApplicationRoleService appRoleService,
-            ApplicationUserManager userManager,
             IApplicationGroupService appGroupService) : base(errorService)
         {
             _appGroupService = appGroupService;
-            _appRoleService = appRoleService;
-            _userManager = userManager;
         }
 
         public ActionResult Index()
         {
             return View();
+        }
+
+        public ActionResult Detail()
+        {
+            return View(GetAll());
         }
 
         public ActionResult ViewAll()
@@ -45,91 +48,124 @@ namespace BTS.Web.Controllers
             return Mapper.Map<IEnumerable<ApplicationGroupViewModel>>(model);
         }
 
-        public ActionResult AddOrEdit(string id = "")
+        [AuthorizeRoles(CommonConstants.System_CanAdd_Role, CommonConstants.System_CanEdit_Role)]
+        public async Task<ActionResult> AddOrEdit(string act, string id = "")
         {
+            ViewBag.action = act;
+
             ApplicationGroupViewModel Item = new ApplicationGroupViewModel();
             if (!string.IsNullOrEmpty(id))
             {
-                var DbItem = _appGroupService.GetDetail(id);
-
-                if (DbItem != null)
+                var DbItem = _appGroupService.GetByID(id);
+                if (DbItem == null)
                 {
-                    Item = Mapper.Map<ApplicationGroupViewModel>(DbItem);
-                    var listRole = _appRoleService.GetListRoleByGroupId(Item.ID);
-                    Item.Roles = Mapper.Map<IEnumerable<ApplicationRole>, IEnumerable<ApplicationRoleViewModel>>(listRole);
+                    return HttpNotFound();
                 }
+                else
+                {
+                    var allRole = RoleManager.Roles.OrderByDescending(x => x.Name); ;
+                    var listRole = _appGroupService.GetRolesByGroupId(id);
 
-                ViewBag.Roles = Mapper.Map<IEnumerable<ApplicationRole>, IEnumerable<ApplicationRoleViewModel>>(_appRoleService.GetAll());
+                    Item = Mapper.Map<ApplicationGroupViewModel>(DbItem);
+                    foreach (var roleItem in allRole)
+                    {
+                        var listItem = new SelectListItem()
+                        {
+                            Text = roleItem.Description,
+                            Value = roleItem.Id,
+                            Selected = listRole.Any(g => g.Id == roleItem.Id)
+                        };
+                        Item.RoleList.Add(listItem);
+                    }
+
+                    var listUser = _appGroupService.GetUsersByGroupId(id);
+                    foreach (var userItem in listUser)
+                    {
+                        var listItem = new SelectListItem()
+                        {
+                            Text = userItem.FullName,
+                            Value = userItem.Id,
+                            Selected = true
+                        };
+                        Item.UserList.Add(listItem);
+                    }
+
+                    return View(Item);
+                }
             }
-            return View(Item);
+            else
+            {
+                var allRole = RoleManager.Roles;
+                foreach (var roleItem in allRole)
+                {
+                    var listItem = new SelectListItem()
+                    {
+                        Text = roleItem.Description,
+                        Value = roleItem.Id,
+                        Selected = false
+                    };
+                    Item.RoleList.Add(listItem);
+                }
+                return View(Item);
+            }
         }
 
         [HttpPost]
-        public async Task<ActionResult> AddOrEdit(ApplicationGroupViewModel Item)
+        [ValidateAntiForgeryToken]
+        [AuthorizeRoles(CommonConstants.System_CanAdd_Role, CommonConstants.System_CanEdit_Role)]
+        public async Task<ActionResult> AddOrEdit(ApplicationGroupViewModel Item, params string[] selectedItems)
         {
             try
             {
-                if (string.IsNullOrEmpty(Item.ID))
+                if (ModelState.IsValid)
                 {
-                    ApplicationGroup newItem = new ApplicationGroup();
-                    newItem.UpdateGroup(Item);
-                    newItem.Name = Item.Name;
-                    newItem.Description = Item.Description;
+                    ApplicationGroup appGroup;
+                    // Create new Group
+                    if (string.IsNullOrEmpty(Item.Id))
+                    {
+                        ApplicationGroup newItem = new ApplicationGroup();
+                        newItem.UpdateApplicationGroup(Item);
+                        appGroup = _appGroupService.Add(newItem);
+                        _appGroupService.Save();
+                        //save group
+                    }
+                    else
+                    {
+                        appGroup = _appGroupService.GetDetail(Item.Id);
 
-                    var appGroup = _appGroupService.Add(newItem);
+                        appGroup.UpdateApplicationGroup(Item);
+                        _appGroupService.Update(appGroup);
+                        _appGroupService.Save();
+                    }
+
+                    _appGroupService.DeleteRolesFromGroup(appGroup.Id);
                     _appGroupService.Save();
-                    //save group
 
                     //add ApplicationRoleGroup
                     var listRoleGroup = new List<ApplicationRoleGroup>();
-                    foreach (var role in Item.Roles)
+                    selectedItems = selectedItems ?? new string[] { };
+                    foreach (var role in selectedItems)
                     {
                         listRoleGroup.Add(new ApplicationRoleGroup()
                         {
-                            GroupId = appGroup.ID,
-                            RoleId = role.ID
+                            GroupId = appGroup.Id,
+                            RoleId = role
                         });
                     }
-                    _appRoleService.AddRolesToGroup(listRoleGroup, appGroup.ID);
-                    _appRoleService.Save();
+                    _appGroupService.AddRolesToGroup(listRoleGroup);
+                    _appGroupService.Save();
 
+                    var users = _appGroupService.GetUsersByGroupId(appGroup.Id);
+                    foreach (var userItem in users)
+                    {
+                        var newUserRoles = _appGroupService.GetLogicRolesByUserId(userItem.Id);
+                        await updateRoles(userItem.Id, newUserRoles);
+                    }
                     return Json(new { success = true, html = GlobalClass.RenderRazorViewToString(this, "ViewAll", GetAll()), message = "Submitted Successfully" }, JsonRequestBehavior.AllowGet);
                 }
                 else
                 {
-                    var appGroup = _appGroupService.GetDetail(Item.ID);
-
-                    appGroup.UpdateApplicationGroup(Item);
-                    _appGroupService.Update(appGroup);
-                    //_appGroupService.Save();
-
-                    //save group
-                    var listRoleGroup = new List<ApplicationRoleGroup>();
-                    foreach (var role in Item.Roles)
-                    {
-                        listRoleGroup.Add(new ApplicationRoleGroup()
-                        {
-                            GroupId = appGroup.ID,
-                            RoleId = role.ID
-                        });
-                    }
-                    _appRoleService.AddRolesToGroup(listRoleGroup, appGroup.ID);
-                    _appRoleService.Save();
-
-                    //add role to user
-                    var listRole = _appRoleService.GetListRoleByGroupId(appGroup.ID);
-                    var listUserInGroup = _appGroupService.GetListUserByGroupId(appGroup.ID);
-                    foreach (var user in listUserInGroup)
-                    {
-                        var listRoleName = listRole.Select(x => x.Name).ToArray();
-                        foreach (var roleName in listRoleName)
-                        {
-                            await _userManager.RemoveFromRoleAsync(user.Id, roleName);
-                            await _userManager.AddToRoleAsync(user.Id, roleName);
-                        }
-                    }
-
-                    return Json(new { success = true, html = GlobalClass.RenderRazorViewToString(this, "ViewAll", GetAll()), message = "Submitted Successfully" }, JsonRequestBehavior.AllowGet);
+                    return Json(new { success = false, message = "Lỗi nhập liệu" }, JsonRequestBehavior.AllowGet);
                 }
             }
             catch (Exception ex)
@@ -138,12 +174,28 @@ namespace BTS.Web.Controllers
             }
         }
 
+        [AuthorizeRoles(CommonConstants.System_CanDelete_Role)]
         public async Task<ActionResult> Delete(string id)
         {
             try
             {
-                var appGroup = _appGroupService.Delete(id);
+                var group = _appGroupService.GetByID(id);
+                if (group == null)
+                {
+                    return HttpNotFound();
+                }
+
+                if (_appGroupService.GetUsersByGroupId(id) != null)
+                {
+                    return Json(new { success = false, message = "Không thể xóa Nhóm còn người dùng" }, JsonRequestBehavior.AllowGet);
+                }
+
+                _appGroupService.DeleteRolesFromGroup(id);
                 _appGroupService.Save();
+
+                _appGroupService.Delete(id);
+                _appGroupService.Save();
+
                 return Json(new { success = true, html = GlobalClass.RenderRazorViewToString(this, "ViewAll", GetAll()), message = "Deleted Successfully" }, JsonRequestBehavior.AllowGet);
             }
             catch (Exception ex)
